% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_all_function.R
\name{Cor_pearson}
\alias{Cor_pearson}
\title{correlation coefficient}
\usage{
Cor_pearson(num, threshold = 0)
}
\arguments{
\item{num}{- a dataframe that all are numeric variables}

\item{threshold}{- a value being coefficient threshold}
}
\value{
a dataframe that contains each pair of column names and corresponding r-square values
}
\description{
the function Cor_pearson() is going to return A data frame that contains each pair of column names in the first column
(name the column “Variable Pairs”) and correlation coefficient (Pearson) for all coefficients whose absolute value
is greater than the correlation threshold (do not repeat any pairs) in the second column
}

