% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_all_function.R
\name{explore}
\alias{explore}
\title{Explore2.0}
\usage{
explore(dafra, switch = "on", threshold = 0, vector = NULL)
}
\arguments{
\item{switch}{A plot switch that can accept three values :off,on,or grid}

\item{threshold}{A threshold cut -of fvalue between 0 and 1 for correlations}

\item{vector}{An optional vector that contains one or more integers and represent the number of bins}
}
\value{
a list
}
\description{
this function is a foundamental explore2.0
dafra- A data frame
}

