% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_all_function.R
\name{is.binary}
\alias{is.binary}
\title{Binary ruler}
\usage{
is.binary(v)
}
\arguments{
\item{v}{-a vector}
}
\value{
True or False
}
\description{
the function is going to check if a vector is binary or not
}

