% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_all_function.R
\name{multiplot}
\alias{multiplot}
\title{Draw different graps in one picture}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\value{
a dataframe that contains each pair of column names and corresponding r-square values
}
\description{
the function is going to draw different graphs in one picture
}

