% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_all_function.R
\name{nonum_table}
\alias{nonum_table}
\title{frequency table for categorical and logical variables}
\usage{
nonum_table(dafra)
}
\arguments{
\item{dafra-a}{dataframe been taken}
}
\value{
a frequent table in list
}
\description{
the function nonum.table() is to return a frequency table for every categorical and logical variable
}

