% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_all_function.R
\name{plot_categ_binary}
\alias{plot_categ_binary}
\title{Gray Bar Graph for categorical and binary variables}
\usage{
plot_categ_binary(data_frame, switch)
}
\arguments{
\item{data_frame-}{a dataframe}

\item{switch-}{a character}
}
\description{
this function is to plot a gray bar graph for every categorical and binary variable if the plot switch parameter is “on” or “grid”.
}

