% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_all_function.R
\name{plot_density_count}
\alias{plot_density_count}
\title{Density-count Histograms}
\usage{
plot_density_count(num, switch = "on", vector = NULL)
}
\arguments{
\item{num}{- a dataframe that all are numeric variables}

\item{switch}{- a character that decide the plot}

\item{vector}{- a vector that decide the bins of histograms}
}
\value{
NUll
}
\description{
the function plot_density_count() is thatIf the plot switch parameter is “on”,then plot a pair of
blue histograms with a vertical red line at the mean (one using counts and the other density) for
every numerical variable at each number of bins integer specified in the bin vector parameter.
If the plot switch is set to “grid”, there should be a grid for each count-bin combination and a
separate grid for each density-bin size combination.
}

