% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_all_function.R
\name{rsquare}
\alias{rsquare}
\title{R-Square for all numeric variables}
\usage{
rsquare(num)
}
\arguments{
\item{num}{- a dataframe that all are numeric variables}
}
\value{
a dataframe that contains each pair of column names and corresponding r-square values
}
\description{
the function rsqure() is going to return a dataframe that contains each pair of column names in the first column (name the column “Variable Pairs”) and the associated r-square value in the second column (name the column “R-Square”)
}

