% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_all_function.R
\name{summary_table}
\alias{summary_table}
\title{Summary for all numeric variables}
\usage{
summary_table(dafra)
}
\arguments{
\item{dafra-a}{dataframe been taken}
}
\value{
a summary statistics table
}
\description{
the function summary.table() is to return a summary statistics table for each numerical variable
}

